refactor data_gen_text ops from basic folder into pdf_qa_gen
